refactor(app): separate actions in different package
