+++ app/src/main/java/fr/carbonit/th/simulation/Adventurer.java
@@ -1,5 +1,7 @@
 package fr.carbonit.th.simulation;
 
+import fr.carbonit.th.simulation.actions.Action;
+
 import java.util.Queue;
 import java.util.UUID;
