+++ app/src/main/java/fr/carbonit/th/simulation/Simulator.java
@@ -1,5 +1,10 @@
 package fr.carbonit.th.simulation;
 
+import fr.carbonit.th.simulation.actions.Action;
+import fr.carbonit.th.simulation.actions.MoveForwardAction;
+import fr.carbonit.th.simulation.actions.TurnLeftAction;
+import fr.carbonit.th.simulation.actions.TurnRightAction;
+
 import java.util.List;
 import java.util.Optional;
