+++ app/src/main/java/fr/carbonit/th/simulation/actions/Action.java
@@ -1,0 +1,11 @@
+package fr.carbonit.th.simulation.actions;
+
+import fr.carbonit.th.simulation.Adventurer;
+
+public abstract class Action implements Comparable<Action> {
+    public abstract void execute(Adventurer adventurer);
+
+    public int compareTo(Action action) {
+        return 1;
+    }
+}
