+++ app/src/main/java/fr/carbonit/th/simulation/actions/MoveForwardAction.java
@@ -1,4 +1,6 @@
-package fr.carbonit.th.simulation;
+package fr.carbonit.th.simulation.actions;
+
+import fr.carbonit.th.simulation.Adventurer;
 
 public class MoveForwardAction extends Action {
