+++ app/src/main/java/fr/carbonit/th/simulation/mappers/AdventurerMapper.java
@@ -1,6 +1,10 @@
 package fr.carbonit.th.simulation.mappers;
 
 import fr.carbonit.th.simulation.*;
+import fr.carbonit.th.simulation.actions.Action;
+import fr.carbonit.th.simulation.actions.MoveForwardAction;
+import fr.carbonit.th.simulation.actions.TurnLeftAction;
+import fr.carbonit.th.simulation.actions.TurnRightAction;
 
 import java.util.LinkedList;
 import java.util.Queue;
